+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for v1.9.7 of SecurityCraft--------------------------
 
+- New: Randomize signature button for the Keycard Reader
 - Fix: Entities around a player sometimes get invisible when the player dismounts a camera
 - Fix: Shields do not take damage when blocking a taser that deals high enough damage
 - Fix: The death message of a player killed by a taser does not contain the player firing the taser